Update composer.json snippet to reference v1.2.*
I just copied the version to my `composer.json` that you had in your `README.md`, but apparently it has some issues with Laravel 4.2, but it appears to be fixed in `1.2.*`. This fix is just to help people like me who are copy-paste happy and don't check the tags right off. =)